+++ mayan/apps/document_downloads/literals.py
@@ -1,11 +1,11 @@
 from django.utils.translation import ugettext_lazy as _
 
-DOCUMENT_FILE_DOWNLOAD_MESSAGE_BODY = _(
-    'The document files have compressed '
+DEFAULT_DOCUMENT_FILE_DOWNLOAD_MESSAGE_BODY = _(
+    'The document files have been compressed '
     'and are available for download using the '
-    'link: %(download_url)s or from '
-    'the downloads area (%(download_list_url)s).'
+    'link: {{ download_url }} or from '
+    'the downloads area ({{ download_list_url }}).'
 )
-DOCUMENT_FILE_DOWNLOAD_MESSAGE_SUBJECT = _(
+DEFAULT_DOCUMENT_FILE_DOWNLOAD_MESSAGE_SUBJECT = _(
     'Document files ready for download.'
 )